+++ app/Functions/components.php
@@ -1,5 +1,10 @@
 <?php
 
+if (!session_id()) {
+    session_save_path(dirname(__DIR__)."/Storage/");
+    session_start();
+}
+
 foreach(glob('../config/*.php') as $configs){
     require_once $configs;
 }
@@ -9,8 +14,4 @@ if (defined('ERRORS_DISPLAY')) {
         ini_set('display_errors', 1);
         error_reporting(E_ALL);
     }
-}
-
-foreach(glob('../routers/*.php') as $routers){
-    require_once $routers;
 }